Changed many things

+++ source/ChunkRenderer.cpp
@@ -7,124 +7,124 @@
 namespace ChunkRenderer {
 
 	void renderPrimitive(QuadPrimitive& p){
-		double color = (double)p.brightness / World::BRIGHTNESSMAX;
+		float color = (float)p.brightness / World::BRIGHTNESSMAX;
 		int x = p.x, y = p.y, z = p.z, length = p.length;
 		ubyte face = 0;
 		if (p.direction == 2) face = 1;
 		else if (p.direction == 3) face = 3;
 		else face = 2;
 #ifdef NERDMODE1
-		Renderer::TexCoord3d(0.0, 0.0, (Textures::getTextureIndex(p.block, face) - 0.5) / 64.0);
+		Renderer::TexCoord3d(0.0f, 0.0f, (Textures::getTextureIndex(p.block, face) - 0.5f) / 64.0);
 		if (p.direction == 0) {
 			if (p.block != Blocks::GLOWSTONE) color *= 0.7;
 			Renderer::Color3d(color, color, color);
-			Renderer::TexCoord2d(0.0, 0.0); Renderer::Vertex3d(x + 0.5, y - 0.5, z - 0.5);
-			Renderer::TexCoord2d(0.0, 1.0); Renderer::Vertex3d(x + 0.5, y + 0.5, z - 0.5);
-			Renderer::TexCoord2d(length + 1.0, 1.0); Renderer::Vertex3d(x + 0.5, y + 0.5, z + length + 0.5);
-			Renderer::TexCoord2d(length + 1.0, 0.0); Renderer::Vertex3d(x + 0.5, y - 0.5, z + length + 0.5);
+			Renderer::TexCoord2d(0.0f, 0.0f); Renderer::Vertex3d(x + 0.5f, y - 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(0.0f, 1.0f); Renderer::Vertex3d(x + 0.5f, y + 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 1.0f); Renderer::Vertex3d(x + 0.5f, y + 0.5f, z + length + 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 0.0f); Renderer::Vertex3d(x + 0.5f, y - 0.5f, z + length + 0.5f);
 		}
 		else if (p.direction == 1) {
 			if (p.block != Blocks::GLOWSTONE) color *= 0.7;
 			Renderer::Color3d(color, color, color);
-			Renderer::TexCoord2d(0.0, 1.0); Renderer::Vertex3d(x - 0.5, y + 0.5, z - 0.5);
-			Renderer::TexCoord2d(0.0, 0.0); Renderer::Vertex3d(x - 0.5, y - 0.5, z - 0.5);
-			Renderer::TexCoord2d(length + 1.0, 0.0); Renderer::Vertex3d(x - 0.5, y - 0.5, z + length + 0.5);
-			Renderer::TexCoord2d(length + 1.0, 1.0); Renderer::Vertex3d(x - 0.5, y + 0.5, z + length + 0.5);
+			Renderer::TexCoord2d(0.0f, 1.0f); Renderer::Vertex3d(x - 0.5f, y + 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(0.0f, 0.0f); Renderer::Vertex3d(x - 0.5f, y - 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 0.0f); Renderer::Vertex3d(x - 0.5f, y - 0.5f, z + length + 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 1.0f); Renderer::Vertex3d(x - 0.5f, y + 0.5f, z + length + 0.5f);
 		}
 		else if (p.direction == 2) {
 			Renderer::Color3d(color, color, color);
-			Renderer::TexCoord2d(0.0, 0.0); Renderer::Vertex3d(x + 0.5, y + 0.5, z - 0.5);
-			Renderer::TexCoord2d(0.0, 1.0); Renderer::Vertex3d(x - 0.5, y + 0.5, z - 0.5);
-			Renderer::TexCoord2d(length + 1.0, 1.0); Renderer::Vertex3d(x - 0.5, y + 0.5, z + length + 0.5);
-			Renderer::TexCoord2d(length + 1.0, 0.0); Renderer::Vertex3d(x + 0.5, y + 0.5, z + length + 0.5);
+			Renderer::TexCoord2d(0.0f, 0.0f); Renderer::Vertex3d(x + 0.5f, y + 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(0.0f, 1.0f); Renderer::Vertex3d(x - 0.5f, y + 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 1.0f); Renderer::Vertex3d(x - 0.5f, y + 0.5f, z + length + 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 0.0f); Renderer::Vertex3d(x + 0.5f, y + 0.5f, z + length + 0.5f);
 		}
 		else if (p.direction == 3) {
 			Renderer::Color3d(color, color, color);
-			Renderer::TexCoord2d(0.0, 0.0); Renderer::Vertex3d(x - 0.5, y - 0.5, z - 0.5);
-			Renderer::TexCoord2d(0.0, 1.0); Renderer::Vertex3d(x + 0.5, y - 0.5, z - 0.5);
-			Renderer::TexCoord2d(length + 1.0, 1.0); Renderer::Vertex3d(x + 0.5, y - 0.5, z + length + 0.5);
-			Renderer::TexCoord2d(length + 1.0, 0.0); Renderer::Vertex3d(x - 0.5, y - 0.5, z + length + 0.5);
+			Renderer::TexCoord2d(0.0f, 0.0f); Renderer::Vertex3d(x - 0.5f, y - 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(0.0f, 1.0f); Renderer::Vertex3d(x + 0.5f, y - 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 1.0f); Renderer::Vertex3d(x + 0.5f, y - 0.5f, z + length + 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 0.0f); Renderer::Vertex3d(x - 0.5f, y - 0.5f, z + length + 0.5f);
 		}
 		else if (p.direction == 4) {
-			if (p.block != Blocks::GLOWSTONE) color *= 0.5;
+			if (p.block != Blocks::GLOWSTONE) color *= 0.5f;
 			Renderer::Color3d(color, color, color);
-			Renderer::TexCoord2d(0.0, 1.0); Renderer::Vertex3d(x - 0.5, y + 0.5, z + 0.5);
-			Renderer::TexCoord2d(0.0, 0.0); Renderer::Vertex3d(x - 0.5, y - 0.5, z + 0.5);
-			Renderer::TexCoord2d(length + 1.0, 0.0); Renderer::Vertex3d(x + length + 0.5, y - 0.5, z + 0.5);
-			Renderer::TexCoord2d(length + 1.0, 1.0); Renderer::Vertex3d(x + length + 0.5, y + 0.5, z + 0.5);
+			Renderer::TexCoord2d(0.0f, 1.0f); Renderer::Vertex3d(x - 0.5f, y + 0.5f, z + 0.5f);
+			Renderer::TexCoord2d(0.0f, 0.0f); Renderer::Vertex3d(x - 0.5f, y - 0.5f, z + 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 0.0f); Renderer::Vertex3d(x + length + 0.5f, y - 0.5f, z + 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 1.0f); Renderer::Vertex3d(x + length + 0.5f, y + 0.5f, z + 0.5f);
 		}
 		else if (p.direction == 5) {
-			if (p.block != Blocks::GLOWSTONE) color *= 0.5;
+			if (p.block != Blocks::GLOWSTONE) color *= 0.5f;
 			Renderer::Color3d(color, color, color);
-			Renderer::TexCoord2d(0.0, 0.0); Renderer::Vertex3d(x - 0.5, y - 0.5, z - 0.5);
-			Renderer::TexCoord2d(0.0, 1.0); Renderer::Vertex3d(x - 0.5, y + 0.5, z - 0.5);
-			Renderer::TexCoord2d(length + 1.0, 1.0); Renderer::Vertex3d(x + length + 0.5, y + 0.5, z - 0.5);
-			Renderer::TexCoord2d(length + 1.0, 0.0); Renderer::Vertex3d(x + length + 0.5, y - 0.5, z - 0.5);
+			Renderer::TexCoord2d(0.0f, 0.0f); Renderer::Vertex3d(x - 0.5f, y - 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(0.0f, 1.0f); Renderer::Vertex3d(x - 0.5f, y + 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 1.0f); Renderer::Vertex3d(x + length + 0.5f, y + 0.5f, z - 0.5f);
+			Renderer::TexCoord2d(length + 1.0f, 0.0f); Renderer::Vertex3d(x + length + 0.5f, y - 0.5f, z - 0.5f);
 		}
 #else
-		float T3d = (Textures::getTextureIndex(p.block, face) - 0.5) / 64.0;
+		float T3d = (Textures::getTextureIndex(p.block, face) - 0.5f) / 64.0;
 		switch (p.direction)
 		{
 			case 0: {
-				if (p.block != Blocks::GLOWSTONE) color *= 0.7;
+				if (p.block != Blocks::GLOWSTONE) color *= 0.7f;
 				float geomentry[] = {
-					0.0, 0.0, T3d, color, color, color, x + 0.5, y - 0.5, z - 0.5,
-					0.0, 1.0, T3d, color, color, color, x + 0.5, y + 0.5, z - 0.5,
-					length + 1.0, 1.0, T3d, color, color, color, x + 0.5, y + 0.5, z + length + 0.5,
-					length + 1.0, 0.0, T3d, color, color, color, x + 0.5, y - 0.5, z + length + 0.5
+					0.0f, 0.0f, T3d, color, color, color, x + 0.5f, y - 0.5f, z - 0.5f,
+					0.0f, 1.0f, T3d, color, color, color, x + 0.5f, y + 0.5f, z - 0.5f,
+					length + 1.0f, 1.0f, T3d, color, color, color, x + 0.5f, y + 0.5f, z + length + 0.5f,
+					length + 1.0f, 0.0f, T3d, color, color, color, x + 0.5f, y - 0.5f, z + length + 0.5f
 				};
 				Renderer::Quad(geomentry);
 			}
 			break;
 			case 1: {
-				if (p.block != Blocks::GLOWSTONE) color *= 0.7;
+				if (p.block != Blocks::GLOWSTONE) color *= 0.7f;
 				float geomentry[] = {
-					0.0, 1.0, T3d, color, color, color, x - 0.5, y + 0.5, z - 0.5,
-					0.0, 0.0, T3d, color, color, color, x - 0.5, y - 0.5, z - 0.5,
-					length + 1.0, 0.0, T3d, color, color, color, x - 0.5, y - 0.5, z + length + 0.5,
-					length + 1.0, 1.0, T3d, color, color, color, x - 0.5, y + 0.5, z + length + 0.5
+					0.0f, 1.0f, T3d, color, color, color, x - 0.5f, y + 0.5f, z - 0.5f,
+					0.0f, 0.0f, T3d, color, color, color, x - 0.5f, y - 0.5f, z - 0.5f,
+					length + 1.0f, 0.0f, T3d, color, color, color, x - 0.5f, y - 0.5f, z + length + 0.5f,
+					length + 1.0f, 1.0f, T3d, color, color, color, x - 0.5f, y + 0.5f, z + length + 0.5f
 				};
 				Renderer::Quad(geomentry);
 			}
 			break;
 			case 2: {
 				float geomentry[] = {
-					0.0, 0.0, T3d, color, color, color, x + 0.5, y + 0.5, z - 0.5,
-					0.0, 1.0, T3d, color, color, color, x - 0.5, y + 0.5, z - 0.5,
-					length + 1.0, 1.0, T3d, color, color, color, x - 0.5, y + 0.5, z + length + 0.5,
-					length + 1.0, 0.0, T3d, color, color, color, x + 0.5, y + 0.5, z + length + 0.5
+					0.0f, 0.0f, T3d, color, color, color, x + 0.5f, y + 0.5f, z - 0.5f,
+					0.0f, 1.0f, T3d, color, color, color, x - 0.5f, y + 0.5f, z - 0.5f,
+					length + 1.0f, 1.0f, T3d, color, color, color, x - 0.5f, y + 0.5f, z + length + 0.5f,
+					length + 1.0f, 0.0f, T3d, color, color, color, x + 0.5f, y + 0.5f, z + length + 0.5f
 				};
 				Renderer::Quad(geomentry);
 			}
 			break;
 			case 3: {
 				float geomentry[] = {
-					0.0, 0.0, T3d, color, color, color, x - 0.5, y - 0.5, z - 0.5,
-					0.0, 1.0, T3d, color, color, color, x + 0.5, y - 0.5, z - 0.5,
-					length + 1.0, 1.0, T3d, color, color, color, x + 0.5, y - 0.5, z + length + 0.5,
-					length + 1.0, 0.0, T3d, color, color, color, x - 0.5, y - 0.5, z + length + 0.5
+					0.0f, 0.0f, T3d, color, color, color, x - 0.5f, y - 0.5f, z - 0.5f,
+					0.0f, 1.0f, T3d, color, color, color, x + 0.5f, y - 0.5f, z - 0.5f,
+					length + 1.0f, 1.0f, T3d, color, color, color, x + 0.5f, y - 0.5f, z + length + 0.5f,
+					length + 1.0f, 0.0f, T3d, color, color, color, x - 0.5f, y - 0.5f, z + length + 0.5f
 				};
 				Renderer::Quad(geomentry);
 			}
 			break;
 			case 4: {
-				if (p.block != Blocks::GLOWSTONE) color *= 0.5;
+				if (p.block != Blocks::GLOWSTONE) color *= 0.5f;
 				float geomentry[] = {
-					0.0, 1.0, T3d, color, color, color, x - 0.5, y + 0.5, z + 0.5,
-					0.0, 0.0, T3d, color, color, color, x - 0.5, y - 0.5, z + 0.5,
-					length + 1.0, 0.0, T3d, color, color, color, x + length + 0.5, y - 0.5, z + 0.5,
-					length + 1.0, 1.0, T3d, color, color, color, x + length + 0.5, y + 0.5, z + 0.5
+					0.0f, 1.0f, T3d, color, color, color, x - 0.5f, y + 0.5f, z + 0.5f,
+					0.0f, 0.0f, T3d, color, color, color, x - 0.5f, y - 0.5f, z + 0.5f,
+					length + 1.0f, 0.0f, T3d, color, color, color, x + length + 0.5f, y - 0.5f, z + 0.5f,
+					length + 1.0f, 1.0f, T3d, color, color, color, x + length + 0.5f, y + 0.5f, z + 0.5f
 				};
 				Renderer::Quad(geomentry);
 			}
 			break;
 			case 5: {
-				if (p.block != Blocks::GLOWSTONE) color *= 0.5;
+				if (p.block != Blocks::GLOWSTONE) color *= 0.5f;
 				float geomentry[] = {
-					0.0, 0.0, T3d, color, color, color, x - 0.5, y - 0.5, z - 0.5,
-					0.0, 1.0, T3d, color, color, color, x - 0.5, y + 0.5, z - 0.5,
-					length + 1.0, 1.0, T3d, color, color, color, x + length + 0.5, y + 0.5, z - 0.5,
-					length + 1.0, 0.0, T3d, color, color, color, x + length + 0.5, y - 0.5, z - 0.5
+					0.0f, 0.0f, T3d, color, color, color, x - 0.5f, y - 0.5f, z - 0.5f,
+					0.0f, 1.0f, T3d, color, color, color, x - 0.5f, y + 0.5f, z - 0.5f,
+					length + 1.0f, 1.0f, T3d, color, color, color, x + length + 0.5f, y + 0.5f, z - 0.5f,
+					length + 1.0f, 0.0f, T3d, color, color, color, x + length + 0.5f, y - 0.5f, z - 0.5f
 				};
 				Renderer::Quad(geomentry);
 			}
